% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_remarkjs_css}
\alias{get_remarkjs_css}
\title{Copy default remark.js stylesheet to a given directory}
\usage{
get_remarkjs_css(to = getwd())
}
\arguments{
\item{to}{where to copy the default stylesheet}
}
\value{
This function returns a logical value indicating whether the operation succeeded or not. Using a missing value for a file or path name will always be regarded as a failure: see \link[base:file.copy]{base::file.copy} for more details.
}
\description{
With \code{get_remarkjs_css()} you can copy the default stylesheet used for the UoL remark.js template to a given directory. This is useful if you want to tune the stylesheet yourself or implement new classes.
}
\examples{
\dontrun{
# Copy the default stylesheet to the current working directory
get_remarkjs_css()

# Copy the default stylesheet to the "temp" directory
get_remarkjs_css(to = "temp")
}

}
