% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{get_uol_palette_manual}
\alias{get_uol_palette_manual}
\title{Manual colours palette generator}
\usage{
get_uol_palette_manual(colours, intensity = c("full", "medium", "light"))
}
\arguments{
\item{colours}{A vector of colour names. See \code{\link{uol_colours}} for a list of possible colours.}

\item{intensity}{Intensity of the desired palette. Possible values are \code{full}, \code{medium}, \code{light}. Defaults to \code{full}.}
}
\value{
A vector of colours, in HEX format.
}
\description{
\code{\link[=get_uol_palette_manual]{get_uol_palette_manual()}} return a vector of colours following the University of Leicester visual identity guidelines. It is possible to select the intensity of the palette, the number of colours, and the method used to pick colours. When a palette of 4-10 colours is requested, the colours in the full palette are sorted by the aforementioned method first, and then the required number of colours are selected by picking equally spaced colours on the ranks scale.
}
\examples{
get_uol_palette_manual(c("red", "blue", "yellow"))
get_uol_palette_manual(c("red", "blue", "yellow"), intensity = "medium")
}
\seealso{
Other get palettes: \code{\link{get_uol_palette_mono}},
  \code{\link{get_uol_palette}}
}
