% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniki.R
\name{miniki}
\alias{miniki}
\title{Minimal-looking light template for a beamer presentation following Karolinska Institutet's visual identity guidelines, with an emphasis on typography}
\usage{
miniki(...)
}
\arguments{
\item{...}{Additional arguments to \code{\link[rmarkdown:beamer_presentation]{rmarkdown::beamer_presentation()}}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}
}
\description{
A beamer template following Karolinska Institutet's visual identity guidelines
}
\examples{

\dontrun{
library(rmarkdown)
draft("my_presentation.Rmd", template = "miniki", package = "templates")
}

}
