% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniuol.R
\name{miniuol}
\alias{miniuol}
\title{Minimal-looking light template for a beamer presentation following the University of Leicester visual identity guidelines, with an emphasis on typography}
\usage{
miniuol(...)
}
\arguments{
\item{...}{Additional arguments to \code{\link[rmarkdown:beamer_presentation]{rmarkdown::beamer_presentation()}}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}
}
\description{
A beamer template following University of Leicester visual identity guidelines
}
\examples{

\dontrun{
library(rmarkdown)
draft("my_presentation.Rmd", template = "miniuol", package = "templates")
}

}
