% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null.R
\name{null}
\alias{null}
\title{A streamlined beamer template based on the default theme}
\usage{
null(...)
}
\arguments{
\item{...}{Additional arguments to \code{\link[rmarkdown:beamer_presentation]{rmarkdown::beamer_presentation()}}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}
}
\description{
A streamlined beamer template absed on the default theme
}
\examples{

\dontrun{
library(rmarkdown)
draft("my_presentation.Rmd", template = "null", package = "templates")
}

}
