% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uol_beamer.R
\name{uol_beamer}
\alias{uol_beamer}
\title{A minimal beamer template following University of leicester visual identity guidelines}
\usage{
uol_beamer(...)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::beamer_presentation}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Write something here...
}
\examples{

\dontrun{
library(rmarkdown)
draft("my_presentation.Rmd", template = "uol_beamer", package = "uolvid")
}

}
