% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uol_beamer_white.R
\name{uol_beamer_white}
\alias{uol_beamer_white}
\title{A minimal, white beamer template following University of Leicester visual identity guidelines}
\usage{
uol_beamer_white(...)
}
\arguments{
\item{...}{Additional arguments to \code{\link[rmarkdown:beamer_presentation]{rmarkdown::beamer_presentation()}}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}
}
\description{
A beamer template following University of Leicester visual identity guidelines
}
\examples{

\dontrun{
library(rmarkdown)
draft("my_presentation.Rmd", template = "uol_beamer_white", package = "plates")
}

}
