% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{uol_colours}
\alias{uol_colours}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 11.}
\usage{
uol_colours
}
\value{
A list of colors, in HEX format.
}
\description{
Use the functions \code{\link{get_uol_palette}}, \code{\link{get_uol_palette_mono}}, \code{\link{get_uol_palette_manual}} to obtain the desired color palette.
Plase note that \code{blue}, \code{grey}, and \code{brown} are all primary colours, while the remaining are secondary ones.
}
\keyword{datasets}
