% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uol_remarkjs.R
\name{uol_remarkjs}
\alias{uol_remarkjs}
\title{A minimal Remark.js template following University of leicester visual identity guidelines}
\usage{
uol_remarkjs(..., css = "default")
}
\arguments{
\item{...}{Additional arguments to \code{xaringan::moon_reader}}

\item{css}{A vector of CSS file paths. A default CSS file is provided in this
package, which was borrowed from \url{https://remarkjs.com}. If the
character vector \code{css} contains the value \code{'default'}, the
default CSS will be used (e.g. \code{css = c('default', 'extra.css')}).}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Write something here...
}
\examples{

\dontrun{
library(rmarkdown)
draft("my_presentation.Rmd", template = "uol_remarkjs", package = "uolvid")
}

}
