% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_linewidth}
\alias{wrap_linewidth}
\title{Wrap output to a given width}
\usage{
wrap_linewidth()
}
\description{
Wrap output to a given width
}
\examples{
\dontrun{
options(width = 60)
knitr::opts_chunk$set(echo = FALSE, fig.align = "center", linewidth = getOption("width"))
uolvid:::wrap_linewidth()
}
}
