% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uol_remarkjs.R
\name{uol_remarkjs}
\alias{uol_remarkjs}
\title{A minimal Remark.js template following University of Leicester visual identity guidelines}
\usage{
uol_remarkjs(..., css = "uol_remarkjs")
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[xaringan:moon_reader]{xaringan::moon_reader()}}}

\item{css}{A vector of CSS file paths. Two default CSS files
(\file{default.css} and \file{default-fonts.css}) are provided in this
package, which was borrowed from \url{https://remarkjs.com}. If the
character vector \code{css} contains a value that does not end with
\code{.css}, it is supposed to be a built-in CSS file in this package,
e.g., for \code{css = c('default', 'extra.css')}), it means
\code{default.css} in this package and a user-provided \code{extra.css}. To
find out all built-in CSS files, use \code{xaringan:::list_css()}.}
}
\description{
This format is a wrapper around \code{\link[xaringan:moon_reader]{xaringan::moon_reader()}}. It includes a custom .css file following visual identity guidelines issued by the University of Leicester.
}
\note{
Do not trust the preview of the presentation from the RStudio viewer or window. Instead, open the presentation with a web browser and refresh to see changes.

Print to \code{.pdf} works best if using Chrome; more details on printing to PDF: \href{https://github.com/yihui/xaringan/issues/65}{#65} and \href{https://github.com/yihui/xaringan/wiki/Export-Slides-to-PDF}{here}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("my_presentation.Rmd", template = "uol_remarkjs", package = "uolvid")
}

}
