% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_linewidth}
\alias{wrap_linewidth}
\title{Wrap output to a given width}
\usage{
wrap_linewidth()
}
\description{
\code{wrap_linewidth()} wrap the output from an R code chunk to a given width, specified by the global option \code{width}. Usage is pretty straighforward: set the global \code{width} option to a given value, set the knitr chunk option \code{linewidth} to \code{getOption("width")}, and call \code{wrap_linewidth()}. Further details \href{https://github.com/yihui/knitr-examples/blob/master/077-wrap-output.Rmd}{here}.
}
\examples{
\dontrun{
options(width = 60)
knitr::opts_chunk$set(echo = FALSE, fig.align = "center", linewidth = getOption("width"))
templates:::wrap_linewidth()
}

}
